put output and analysis to other files
